sim directory cleanups: - dir names "_sim" drop - import removals - adam_logger -> logger - etc
