+++ sim/functional/allocator/.gitignore
