+++ sim/functional/allocator/utils
