+++ sim/functional/circ_fifo/Makefile
