+++ sim/functional/circ_fifo/utils
