+++ sim/functional/mesh_wh_xy_noc/test.py
@@ -1,9 +1,9 @@
 from logging import INFO, DEBUG
 from cocotb.regression import TestFactory
-from cocotb.triggers import RisingEdge, FallingEdge, ReadOnly, ClockCycles, Combine
+from cocotb.triggers import ClockCycles
 from cocotb.result import TestSuccess
 import os
-import numpy as np
+from random import randint
 from utils.functions import *
 from noc import WHNoCTB
 from noc import NoCMetrics
