+++ sim/functional/mesh_wormhole_switch/gtk.sh
