--- sim/functional/sample_test/.gitignore
@@ -1,5 +1,0 @@
-sim_build
-__pycache__
-dump.vcd
-results.xml
-.idea/
