+++ sim/functional/utils/traffic_pattern.py
@@ -247,54 +247,3 @@ class TrafficPattern(object):
 
 
 tp = TrafficPattern([5,5])
-# # print(tp.u_rand(0))
-# node = NodeAddr([0, 1], tp.dim)
-# print(tp.locality(node.id).id)
-#
-# node = NodeAddr([0, 0], tp.dim)
-# print(tp.locality(node.id).id)
-#
-#
-# node = NodeAddr([1, 1], tp.dim)
-# print(tp.locality(node.id).id)
-#
-# node = NodeAddr([2, 2], tp.dim)
-# print(tp.locality(node.id).id)
-
-# node = NodeAddr([2, 2], tp.dim)
-# print(tp.nearest_neighbor(node.id, radius=2, nearest_prob=.6).addr)
-# print(tp.nearest_neighbor(node.id, radius=1, nearest_prob=0.55).addr)
-# print(tp.nearest_neighbor(node.id, radius=1, nearest_prob=0.7).addr)
-
-# node = NodeAddr([2, 2], tp.dim)
-# print(tp.hotspot(node.id).id)
-# print(tp.nearest_neighbor(node.id, radius=1, nearest_prob=0.55).addr)
-# print(tp.nearest_neighbor(node.id, radius=1, nearest_prob=0.7).addr)
-
-
-
-# print("INPUT NODE")
-
-# print(f"coords {node.row}, {node.col}")
-# print(f"id     {node.id}")
-# print("")
-#
-# new_node = tp.bit_permutation(node, "invert")
-# print(f"coords {new_node.row}, {new_node.col}")
-# print(f"id     {new_node.id}")
-#
-# new_node = tp.bit_permutation(node, "reverse")
-# print(f"coords {new_node.row}, {new_node.col}")
-# print(f"id     {new_node.id}")
-#
-# new_node = tp.bit_permutation(node, "rotate")
-# print(f"coords {new_node.row}, {new_node.col}")
-# print(f"id     {new_node.id}")
-#
-# new_node = tp.bit_permutation(node, "shuffle")
-# print(f"coords {new_node.row}, {new_node.col}")
-# print(f"id     {new_node.id}")
-#
-# new_node = tp.digit_permutation(node, "neighbour")
-# print(f"coords {new_node.row}, {new_node.col}")
-# print(f"id     {new_node.id}")
