+++ sim/functional/virtual_channel/utils
