+++ synth/synth_noc_comb.py
@@ -1,10 +1,7 @@
 import subprocess
 import os
-import time
-import argparse
 from utils.rav import *
-from utils.adam_logger import *
-import glob
+from utils.logger import *
 import sys
 import numpy as np
